MDEV-35658 Assertion `commit_trx' failed in test galera_as_master
The test issues a simple INSERT statement, while sql_log_bin = 0.
This option disables writes to binlog.  However, since MDEV-7205,
the option does not affect Galera, so changes are still replicated.
So sql_log_bin=off, "partially" disabled the binlog and the INSERT
will involve both binlog and innodb, thus requiring internal 2 phase
commit (2PC). In 2PC INSERT is first prepared, which will make it
transition to PREPARED state in innodb, and later committed which
causes the new assertion from MDEV-24035 to fail.
Running the same test with sql_log_bin enabled also results in 2PC,
but the execution has one more step for ordered commit, between prepare
and commit. Ordered commit causes the transaction state to transition
back to TRX_STATE_NOT_STARTED. Thus avoiding the assertion.
To fix this, the patch makes sure that when binlog is disabled
we switch wsrep_binlog_emulation on (and viceversa). So that binlog
is not involved, and therefore skipping the internal 2PC.