MDEV-36150 Retain original Xid_log_event in Galera Cluster
Prior to this change, Galera overwrote xid used to prepare
a transaction with a xid containing wsrep seqno and UUID.
This however is not fully compatible with partition engine
and atomic DDL recovery.

In order to avoid overwriting xid during prepare, store the
wsrep transaction seqno and UUID into Xid_log_event if the
transaction is wsrep transaction. This way the seqno and UUID
will be available to reconstruct the wsrep XID for commit
during binlog coordinated recovery.

This changes on-disk format of Xid_log_event to contain
additional 24 bytes for wsrep seqno and UUID if the
transaction is wsrep transaction. As Xid_log_event has no
meaning in replication and used only in recovery, the change
is backwards compatible.

The Xid_log_event output in mariadb-binlog is extended to
print also wsrep_seqno and wsrep_uuid if they are set.