Fix wsrep_xid use in lock_wait_wsrep_kill()
In lock_wait_wsrep_kill() the assigned xid was checked whether
it is wsrep XID if the victim transaction was found in prepared
state. This condition is now never fulfilled because wsrep XID
is not used for 2PC. Let `wsrep_thd_bf_abort()` decide whether
the victim should be BF aborted.